% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{ody_add_tbl_violindotbox}
\alias{ody_add_tbl_violindotbox}
\title{Add descriptive table to a "Violindotbox" plot}
\usage{
ody_add_tbl_violindotbox(p, n_dec = 1, add_test = FALSE)
}
\arguments{
\item{p}{A ggplot object generated by \code{ody_plot_violindotbox}.}

\item{n_dec}{Number of decimals to show in the descriptive table.}

\item{add_test}{Logical indicating if a statistical test should be added to the table.}
}
\value{
A patchwork object that combines the original plot and the descriptive table.
}
\description{
Add descriptive table to a "Violindotbox" plot
}
