% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_utils.R
\name{ody_add_version}
\alias{ody_add_version}
\title{Get a version date for a render.}
\usage{
ody_add_version(file_name, extension = "html", path = ".")
}
\arguments{
\item{file_name}{File name.}

\item{extension}{File extension.}

\item{path}{The location the file will be stored}
}
\value{
A character
}
\description{
Get a version date for a render.
}
\details{
The function looks for the possible files named as file_name followed by today's date in path. If none, the current date is returned as version indicator. If any, a number is added to the date.
}
