% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_utils.R
\name{ody_change_names}
\alias{ody_change_names}
\title{Change column names}
\usage{
ody_change_names(data_frame, names_df)
}
\arguments{
\item{data_frame}{Data frame the names should by modified}

\item{names_df}{2 columns data frame. The first column must contain the new names and the second one, the current ones.}
}
\value{
data_frame with names changes according to names_df.
}
\description{
Modify column names according to specified changes in a variables data frame.
}
