% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpt.R
\name{ody_correct}
\alias{ody_correct}
\title{Text corrector}
\usage{
ody_correct(prompt)
}
\arguments{
\item{prompt}{Text to correct}
}
\value{
Corrected text
}
\description{
Simple shortcut to use OdyGPT as a text corrector
}
