% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_wrangling.R
\name{ody_define_timepoints}
\alias{ody_define_timepoints}
\title{Define the extraction timepoint from a sequence of events}
\usage{
ody_define_timepoints(extractions, events)
}
\arguments{
\item{extractions}{Extractions data frame:
\itemize{
\item Each row represents an extraction.
\item The first column must be the patient id used to join with the events data frame.
\item The extraction date is assumed to be the first date column.
}}

\item{events}{Events data frame. Each row represents the history of a single patient and the first column must be the patient id used to join with the extractions data frame.}
}
\value{
A tibble
}
\description{
Define the extraction timepoint from a sequence of events
}
