% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_utils.R
\name{ody_exofilter}
\alias{ody_exofilter}
\title{External Filter Function}
\usage{
ody_exofilter(external_df, id_col)
}
\arguments{
\item{external_df}{A data frame containing reference data for filtering.}

\item{id_col}{The column name in both data frames that identifies subjects.}
}
\value{
A function that takes a data frame and a filter expression, returning the filtered data frame.
}
\description{
Creates a filtering function that selects rows from a data frame based on an external reference data frame and a filter expression.
}
