% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{ody_extract_km}
\alias{ody_extract_km}
\title{Extract KM table from a survfit object}
\usage{
ody_extract_km(km_fit, gt_table = FALSE, n_dec = 1)
}
\arguments{
\item{km_fit}{KM model}

\item{gt_table}{logic, returmn the table as a gt?}

\item{n_dec}{Number of decimals}
}
\description{
Extract KM table from a survfit object
}
