% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_extraction.R
\name{ody_extract_vhlab_pdf}
\alias{ody_extract_vhlab_pdf}
\title{Extract Vall d'Hebron lab data from a PDF}
\usage{
ody_extract_vhlab_pdf(pdf_path, write_xlsx = TRUE)
}
\arguments{
\item{pdf_path}{Path to the PDF file.}

\item{write_xlsx}{Logical indicating if the extracted data should be written
to an xlsx file}
}
\value{
A tibble containing the extracted and processed lab data.
}
\description{
This function extracts the lab data from a PDF file generated by the Vall
d'Hebron
}
