% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_quality.R
\name{ody_filter_fails}
\alias{ody_filter_fails}
\title{Filter Fails}
\usage{
ody_filter_fails(tbl)
}
\arguments{
\item{tbl}{A tibble that contains an \code{.ok} column for the filtering criteria.}
}
\value{
A filtered tibble without the \code{.ok} column.
}
\description{
Filters a tibble to only include cases that are not okay or have missing values
in the \code{.ok} column. Removes the \code{.ok} column from the result.
}
\details{
The main purpose of this function is to filter cases that do not meet
a given verification criterion in the quality check process of a REDCap project.
}
