% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_utils.R
\name{ody_glue2lang}
\alias{ody_glue2lang}
\title{Glue to Languange}
\usage{
ody_glue2lang(..., .envir = parent.frame(), .eval = FALSE)
}
\arguments{
\item{...}{Expressions to glue and transform into language objects.}

\item{.envir}{Environment to evaluate each expression in.}

\item{.eval}{If TRUE, the function evaluates the language object.}
}
\value{
A language object or the its evaluation if .eval = TRUE
}
\description{
Glue a string and convert it to a language object.
}
