% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpt.R
\name{ody_gpt}
\alias{ody_gpt}
\title{Ask to OdyGPT}
\usage{
ody_gpt(prompt)
}
\arguments{
\item{prompt}{Text with your question}
}
\description{
Ask to OdyGPT
}
\details{
A valid OPEN_API_kEY stored in .Renviron is required.
}
