% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_utils.R
\name{ody_gt2image}
\alias{ody_gt2image}
\title{Convert GT Table to Image}
\usage{
ody_gt2image(gt_table, type = c("raster", "ggplot"), zoom = 2)
}
\arguments{
\item{gt_table}{The GT table object to be converted into an image.}

\item{type}{The type of output image. Either "raster" for a raster image or "ggplot"
for a plot created with ggplot. Defaults to "raster".}

\item{zoom}{Zoom factor for the GT table rendering, where higher values result in
higher resolution images. Defaults to 2.}
}
\value{
An image object, either of class \code{magick-image} (for "raster" type) or
a ggplot object (for "ggplot" type).
}
\description{
This function converts a GT table object into an image file. It supports outputting
the image either as a raster image directly or as a plot made with ggplot.
}
\details{
"raster" output can be used with ggplot2::annotate_raster to add the image on a ggplot. "ggplot" output is usefull in combination with "patchwork".
}
