% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_utils.R
\name{ody_label_df}
\alias{ody_label_df}
\title{Label Data Frame}
\usage{
ody_label_df(raw_data, dictionary)
}
\arguments{
\item{raw_data}{A data frame containing the raw data to be labeled.}

\item{dictionary}{A dictionary data frame with the following columns:
\itemize{
\item \code{variable}: The name of the variable to be labeled.
\item \code{variable_label} (optional): The label to assign to the variable.
\item \code{value}: The value to be labeled.
\item \code{value_label}: The label to assign to the value.
Note that columns \code{variable} and \code{variable_label} will contain as many repeated values as the number of values in the \code{value} column.
}}
}
\value{
A data frame with labeled variables and corresponding values according to the provided dictionary.
}
\description{
This function labels a data frame according to a provided dictionary. It sets variable labels and value labels for specified variables within the data frame.
}
