% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aes.R
\name{ody_make_ae_gt}
\alias{ody_make_ae_gt}
\title{Make a gt table of Adverse Events}
\usage{
ody_make_ae_gt(ae_tbl, total_n = NULL, spanner_names = NULL, dec = 1)
}
\arguments{
\item{ae_tbl}{Adverse Events table. The output of ody_make_ae_tbl.}

\item{total_n}{Optional, if presents, a percentage is added to the count}

\item{spanner_names}{Optional, if ae_tbl is a list, the names of the spanners on the table.}

\item{dec}{Number of decimals of the percentages.}
}
\value{
A gt table
}
\description{
Make a gt table of Adverse Events
}
