% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aes.R
\name{ody_make_ae_tbl}
\alias{ody_make_ae_tbl}
\title{Make an Adverse Events Table}
\usage{
ody_make_ae_tbl(
  ae_data,
  pac_id,
  grade_num,
  grade_fct,
  term,
  upper_term = NULL,
  group_fct = NULL,
  add_overall = FALSE
)
}
\arguments{
\item{ae_data}{Data frame of Adverse Events}

\item{pac_id}{Patients identrifier column in ae_data}

\item{grade_num}{Numeric column indicating the grade of each AE}

\item{grade_fct}{Factor column indicating how to group the AEs in the output table.}

\item{term}{Charachter column with the AE name.}

\item{upper_term}{Optional, a higher level name.}

\item{group_fct}{Optional, additional grouping factor.}

\item{add_overall}{If there is a group_fct, add_overall = TRUE adds an overall count to the by group_fct list.}
}
\value{
A tibble if group_fct = NULL. Otherwise, a list of tibbles
}
\description{
Make an Adverse Events Table
}
