% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdx.R
\name{ody_pdx_model_percentage_response}
\alias{ody_pdx_model_percentage_response}
\title{Binomial GLMM to test PDX responsiveness}
\usage{
ody_pdx_model_percentage_response(data_frame)
}
\arguments{
\item{data_frame}{A data frame where:
\itemize{
\item The first column contains the PDX id's.
\item The second column is a two levels factors for sensitivity.
\item The third column is a two levels factor for treatment.
\item The fourth column is the response variable (integer from 0 to 100).
}}
}
\value{
A HTML report in the working directory.
}
\description{
Runs a binomial Genralized Mixed Model to test wether the response to a treatment depends on the sensitivity. The response variable is assumed to be a count of "positive" findings over a total count of 100.
}
