% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{ody_plot_violindotbox}
\alias{ody_plot_violindotbox}
\title{Plot Dots, Boxplot, and Violin Plot}
\usage{
ody_plot_violindotbox(
  data,
  x,
  y,
  no_violin = FALSE,
  compare = FALSE,
  p_adj = "fdr",
  brackets_pos = 1.05,
  ...
)
}
\arguments{
\item{data}{A data frame containing the variables for plotting.}

\item{x}{A string representing the name of the x variable (categorical).}

\item{y}{A string representing the name of the y variable (numeric).}

\item{no_violin}{Logical indicating if the half violin plot should be removed.}

\item{compare}{Logical indicating if a statistical comparison should be added.}

\item{p_adj}{A string indicating the method to adjust the p-values. See \link[stats]{p.adjust}.}

\item{brackets_pos}{A numeric value indicating the relative position of the brackets that represent the statistical comparison.}

\item{...}{Additional arguments to be passed to \link[ggpubr]{compare_means} (the function
used to compare the means) and to \link[ggpubr]{geom_bracket} (the function used to add the brackets).}
}
\value{
A ggplot object that can be further customized.
}
\description{
This function generates a composite plot that includes a half violin plot, a half dot plot,
and a boxplot. It also adds statistical comparisons between groups.
}
