% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap.R
\name{ody_rc_add_import_date}
\alias{ody_rc_add_import_date}
\title{Add the redcap_data import date to a file name}
\usage{
ody_rc_add_import_date(file_name, extension = "csv")
}
\arguments{
\item{file_name}{The name of the file to be saved.}

\item{extension}{The extension of the file. Default is "csv".}
}
\value{
A string with the new file name.
}
\description{
Add the redcap_data import date to a file name
}
