% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap.R
\name{ody_rc_clean_missing_codes}
\alias{ody_rc_clean_missing_codes}
\title{Clean overlapping user-missing codes in labelled variables}
\usage{
ody_rc_clean_missing_codes(rc_df)
}
\arguments{
\item{rc_df}{A data frame (typically obtined from \code{ody_rc_select}) with
labelled variables.}
}
\value{
A data frame with updated user-defined missing codes.
}
\description{
Removes user-defined missing codes (\code{labelled::na_values}) that overlap with
existing value labels (\code{labelled::val_labels}) in each column of a data
frame. This helps avoid conflicts before formatting or converting labelled
variables.
}
\details{
For each column, if any value in \code{na_values} is also present among
\code{val_labels}, the overlapping values are removed from \code{na_values}.
Informative messages are emitted indicating which codes were removed and from
which variable.
}
