% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap.R
\name{ody_rc_completeness}
\alias{ody_rc_completeness}
\title{Verify Completeness of a RedCap-derived dataframe}
\usage{
ody_rc_completeness(
  data_frame,
  id_var = {
     attr(get("redcap_data"), "id_var")
 },
  count_user_na = FALSE,
  conditions_list = "from_metadata",
  metadata = {
     attr(get("redcap_data"), "metadata")
 },
  missing_codes = {
     attr(get("redcap_data"), "missing")
 },
  report = TRUE
)
}
\arguments{
\item{data_frame}{The data frame to verify.}

\item{id_var}{Variable used as ID. By default the functions looks at the "id_var" attribute of redcap_data.}

\item{count_user_na}{If FALSE (the default) only regular missing values are count as missing values. Set this argument to TRUE to also take into account the explicit user defined missing values.}

\item{conditions_list}{A list to define the conditional presence of the variables. Needed when the presence of a variable depends on the value of other variable. By default the list is based on the branching logic defined in the metadata.}

\item{metadata}{Used metadata if conditions_list = "from_metadata". By default, the function gets it from the attribute "metadata" of the redcap_data.}

\item{missing_codes}{Redcap_data missing codes. Only needed if conditions_list is based on the metadata. By default is the attribute "missing_codes".}

\item{report}{Render a report or output a tibble?}
}
\value{
An html report or a tibble
}
\description{
Verify Completeness of a RedCap-derived dataframe
}
