% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap_project.R
\name{ody_rc_current}
\alias{ody_rc_current}
\title{Get the name and current import of the project}
\usage{
ody_rc_current(as_list = FALSE)
}
\arguments{
\item{as_list}{If FALSE, the information is printed on the console. Otherwise, it returns a list (usefull for reports)}
}
\description{
Get the name and current import of the project
}
\details{
"Where am I?" addin calls to this function
}
