% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap.R
\name{ody_rc_filter_subject}
\alias{ody_rc_filter_subject}
\title{Filter a RedCap import by subject id}
\usage{
ody_rc_filter_subject(redcap_data, subjects_vector)
}
\arguments{
\item{redcap_data}{RedCap data import.}

\item{subjects_vector}{Vector with the subjects id's to be kept.}
}
\value{
A filtered RedCap import.
}
\description{
Filter a RedCap import by subject id
}
