% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap.R
\name{ody_rc_format}
\alias{ody_rc_format}
\title{Format RedCap variables}
\usage{
ody_rc_format(rc_df, keep_user_na = FALSE, clean_missing_codes = FALSE)
}
\arguments{
\item{rc_df}{Dataframe derived from a RedCap import with ody_rc_select}

\item{keep_user_na}{Logical. Should user defined missing values be kept? Default is FALSE, so when formatting, user defined missing values are replaced by regular NAs.}

\item{clean_missing_codes}{Logical. Should user defined missing codes that
overlap with value labels be cleaned before formatting? Default is FALSE.
This is usefull in those rare occasions when the label of a variable value is
equal to one of the user defined missing codes (e.g "OTH" in the UNLOCK
project is a missing data code and also a data label in tum_histology)}
}
\value{
A tibble
}
\description{
Format all variables from an ody_rc_select dataframe according to the
variable metadata.
}
\details{
Formatting is performed as follows:
\itemize{
\item Values defined as numeric in redcap -> as.numeric (also
redcap_repeat_instance).
\item Values defined as date in redcap -> ymd
\item Values defined as datetime in redcap -> ymd_hm
\item Values labelled in redcap -> to_factor
\item Other -> as.character
\item User defined missing values -> NA
}
}
