% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap.R
\name{ody_rc_get_metadata}
\alias{ody_rc_get_metadata}
\title{Get REDCap Metadata}
\usage{
ody_rc_get_metadata(token, url = "https://redcap.vhio.net/redcap/api/")
}
\arguments{
\item{token}{A REDCap API token. If NULL, the user will be prompted to enter it.}

\item{url}{The API URL for the REDCap instance. Defaults to \code{"https://redcap.vhio.net/redcap/api/"}.}
}
\value{
The metadata of the RedCap project.
}
\description{
This function retrieves metadata from a RedCap project.
}
