% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap.R
\name{ody_rc_import}
\alias{ody_rc_import}
\title{Import a RedCap Proyect}
\usage{
ody_rc_import(
  token = NULL,
  url = "https://redcap.vhio.net/redcap/api/",
  label = TRUE,
  nest = TRUE
)
}
\arguments{
\item{token}{Hi}

\item{url}{Hi}

\item{label}{Hi}

\item{nest}{Hi}
}
\value{
A nested tibble
}
\description{
Import a RedCap Proyect
}
