% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap_project.R
\name{ody_rc_init}
\alias{ody_rc_init}
\title{Start a RedCap project in RStudio}
\usage{
ody_rc_init(token = NULL, url = "https://redcap.vhio.net/redcap/api/")
}
\arguments{
\item{token}{Project token. If not provided, a dialog promp will ask for it}

\item{url}{URL of the RedCap server (VHIO server by default).}
}
\description{
Start a RedCap project in RStudio
}
