% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap.R
\name{ody_rc_kill_viewers}
\alias{ody_rc_kill_viewers}
\title{Kill all RedCap Viewer processes}
\usage{
ody_rc_kill_viewers()
}
\description{
This function searches for all viewer processes launched by \code{ody_rc_view()}
in the caller environment (typically the global environment) and kills them.
It also removes the corresponding process objects from the environment.
}
