% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap_project.R
\name{ody_rc_refresh_datasets}
\alias{ody_rc_refresh_datasets}
\title{Refresh the dasest list}
\usage{
ody_rc_refresh_datasets(rc_data = redcap_data)
}
\arguments{
\item{rc_data}{RedCap data}
}
\description{
Refresh the dasest list
}
