% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap.R
\name{ody_rc_simplify_selection}
\alias{ody_rc_simplify_selection}
\title{Simplify a selection by dropping RedCap structure variables}
\usage{
ody_rc_simplify_selection(selected_data, join = FALSE)
}
\arguments{
\item{selected_data}{A data frame or list of data frames (typically the
output of ody_rc_select or ody_rc_select_form).}

\item{join}{Logical. If TRUE and selected_data is a list of data frames,
the simplified data frames will be joined with \code{dplyr::full_join} by all
the common columns.}
}
\value{
A simplified data frame or list of data frames without the specified
RedCap structural columns.
}
\description{
Remove RedCap structural columns (redcap_event_name, redcap_form_name,
redcap_instance_type, redcap_instance_number) from a data frame or list of
data frames.
}
