% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap_project.R
\name{ody_rc_timetravel}
\alias{ody_rc_timetravel}
\title{Paradox-Free Time Travelling}
\usage{
ody_rc_timetravel(timepoint)
}
\arguments{
\item{timepoint}{Timepoint pattern.}
}
\description{
Load previous imports and datasets.
}
