% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap.R
\name{ody_rc_translate_meddra}
\alias{ody_rc_translate_meddra}
\title{Translate MedDRA Codes}
\usage{
ody_rc_translate_meddra(rc_df)
}
\arguments{
\item{rc_df}{A dataframe obtained from a RedCap import using the function \code{ody_rc_select}.}
}
\value{
A dataframe with MedDRA codes substituted by their corresponding labels.
}
\description{
Replace MedDRA codes in the dataframe with their corresponding descriptions.
}
