% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_utils.R
\name{ody_read_data}
\alias{ody_read_data}
\title{Read a data file from the project's data folder}
\usage{
ody_read_data(data_file, sheet = NULL, guess_cols = FALSE, ...)
}
\arguments{
\item{data_file}{Name of the data file to read.}

\item{sheet}{Optional; if the file is an Excel file, specify the sheet to read.}

\item{guess_cols}{Logical; if TRUE, column types are guessed, otherwise
all columns are read as text.}

\item{...}{Additional arguments passed to the underlying reading functions
(\code{readxl::read_excel} for Excel files and \code{vroom::vroom} for other files).}
}
\value{
A data frame containing the contents of the specified file.
}
\description{
Read a data file from the project's data folder
}
\details{
This function is intended to be used within "odytools" projects.
It assumes data files are located in a subdirectory named "data"
(inside the project directory) and the function searches for files
within that folder and its subfolders.
}
\examples{
\dontrun{
df <- ody_read_data("mydata.csv")
df <- ody_read_data("mydata.xlsx", guess_cols = TRUE)
}
}
