% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_quality.R
\name{ody_render_quality_report}
\alias{ody_render_quality_report}
\title{Render a quality report}
\usage{
ody_render_quality_report(
  data_frame,
  project_name = "project",
  missing_values = "",
  id_var = "row_number",
  conditions_list = "no",
  add_data = FALSE,
  max_integer_distinct = 10,
  output_dir = getwd()
)
}
\arguments{
\item{data_frame}{data frame to report.}

\item{project_name}{Name to use in the file name.}

\item{missing_values}{vector with the values (apart from the regular NA and "") used as missing. Default is NULL.}

\item{id_var}{ID variable.}

\item{conditions_list}{List of conditions used as argument of verify_completeness.}

\item{add_data}{If TRUE, raw data is added to the report.}

\item{max_integer_distinct}{ody_verify_conformance argument.}

\item{output_dir}{Location to save the report.}
}
\value{
A quality report HTML.
}
\description{
Render a quality report
}
