% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_utils.R
\name{ody_repair_dates}
\alias{ody_repair_dates}
\title{Repair date-like columns into Date class}
\usage{
ody_repair_dates(data, ...)
}
\arguments{
\item{data}{A data frame.}

\item{...}{One or more columns to repair.}
}
\value{
A data frame with the selected columns converted to Date.
}
\description{
Convert messy date-like columns (character or numeric) into Date objects.
}
\details{
The function checks each selected value and attempts the following conversions, in order:
\itemize{
\item If NA, leaves as NA.
\item If lubridate::is.timepoint(), converts to Date.
\item If a 5-digit number (Excel serial), converts using janitor::excel_numeric_to_date().
\item If  matches "2 digits - 2 digits - 4 digits", parses with lubridate::dmy().
\item If matches "4 digits - 2 digits - 2 digits", parses with lubridate::ymd().
\item Otherwise the value becomes NA.
}
}
