% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_quality.R
\name{ody_report_fails}
\alias{ody_report_fails}
\title{Report Fails}
\usage{
ody_report_fails(tbl)
}
\arguments{
\item{tbl}{A tibble containing verification results with a failing criterion.}
}
\value{
A tibble with a message if no fails are detected, or a report table of the failing cases.
}
\description{
Generates a gt table of the cases that have failed verification.
}
\details{
The main purpose of this function is to report the output table of \code{ody_filter_fails}.
}
