% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_utils.R
\name{ody_save_path}
\alias{ody_save_path}
\title{Generate a file path for saving output with versioning.}
\usage{
ody_save_path(...)
}
\arguments{
\item{...}{Components of the path and file name.}
}
\value{
A character string representing the save path.
}
\description{
This function constructs a save path using the provided components and appends a version (today's date) to the file name.
}
