% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_tables.R
\name{ody_summarise_df}
\alias{ody_summarise_df}
\title{ody_summarise_df}
\usage{
ody_summarise_df(
  data_frame,
  grouping_var = NULL,
  conditions_list = NULL,
  exclude = NULL,
  use_NA = c("no", "ifany", "always"),
  num_dec = 3,
  prop_dec = 3
)
}
\arguments{
\item{data_frame}{data frame}

\item{grouping_var}{A grouping variable}

\item{conditions_list}{Conditions list to filter variables}

\item{exclude}{Variables that should be excluded}

\item{use_NA}{Add missing values in the report?}

\item{num_dec}{Number of decimals in the summary of a numeric variable}

\item{prop_dec}{Number of decimals in the proportions}
}
\value{
A list
}
\description{
ody_summarise_df
}
