% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_utils.R
\name{ody_timetravel}
\alias{ody_timetravel}
\title{Paradox-Free Time Travelling}
\usage{
ody_timetravel(timepoint)
}
\arguments{
\item{timepoint}{Timepoint pattern.}
}
\description{
Replace the current datasets list with the one from a previous backup stored in the data/backups folder.
}
\details{
The back-ups are named after the project and the back-up date. The timepoint pattern is a regular expression to match the name of the back-up file. The pattern must match one and only one back-up file.
}
