% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_quality.R
\name{ody_verify_completeness}
\alias{ody_verify_completeness}
\title{Verify Completeness}
\usage{
ody_verify_completeness(
  data_frame,
  missing_values = NULL,
  conditions_list = NULL,
  id_var = "row_number"
)
}
\arguments{
\item{data_frame}{The data frame to be checked.}

\item{missing_values}{Values considered as missing. Actual NA's and empty characters ("") are always considered missing values.}

\item{conditions_list}{A list to define the conditional presence of the variables. Needed when the presence of a variable depends on the value of other variable.}

\item{id_var}{The id variable. By default, the function uses row numbers.}
}
\value{
A completeness tibble.
}
\description{
Verifies the completeness of each variable of a data frame.
}
