% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_quality.R
\name{ody_verify_conformance}
\alias{ody_verify_conformance}
\title{Verify Conformance}
\usage{
ody_verify_conformance(
  data_frame,
  missing_values = NULL,
  max_integer_distinct = 10
)
}
\arguments{
\item{data_frame}{The data frame to be checked.}

\item{missing_values}{Values considered as missing. Actual NA's and empty characters ("") are always considered missing values.}

\item{max_integer_distinct}{Max number of different integer values that will be reported individually.}
}
\value{
A conformance tibble.
}
\description{
Cheks the format of the variables of a data frame.
}
