% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap_methods.R
\name{print.odytools_dataset}
\alias{print.odytools_dataset}
\title{Print a RedCap dataset}
\usage{
\method{print}{odytools_dataset}(x, ...)
}
\arguments{
\item{x}{RedCap dataset created inside an Odytools REDCap project.}

\item{...}{Additional arguments (added for compatibility with base::print).}
}
\description{
Print a RedCap dataset
}
