% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap_methods.R
\name{print.odytools_datasets_list}
\alias{print.odytools_datasets_list}
\title{Print a RedCap dataset list}
\usage{
\method{print}{odytools_datasets_list}(x, ...)
}
\arguments{
\item{x}{RedCap dataset list created inside an Odytools REDCap project.}

\item{...}{Additional arguments (added for compatibility with base::print).}
}
\description{
Print a RedCap dataset list
}
