% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap_methods.R
\name{print.odytools_redcap}
\alias{print.odytools_redcap}
\title{Print a RedCap import}
\usage{
\method{print}{odytools_redcap}(x, ...)
}
\arguments{
\item{x}{RedCap data imported with ody_rc_import.}

\item{...}{Additional arguments (added for compatibility with base::print).}
}
\description{
Print a RedCap import
}
