% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap_methods.R
\name{select.odytools_redcap}
\alias{select.odytools_redcap}
\title{Select variables from a RedCap import}
\usage{
\method{select}{odytools_redcap}(.data, ...)
}
\arguments{
\item{.data}{RedCap data imported with ody_rc_import.}

\item{...}{Variable names to select. If the name of a form is provided, all the variables belonging to that form will be selected.}
}
\value{
A tibble with the selected variables.
}
\description{
Select variables from a RedCap import
}
\details{
The funtion first tries to use dplyr::select.tbl_df to select the variables. If the variables are not found, the function uses ody_rc_select to extract the variables from the RedCap import.
See \link[odytools]{ody_rc_select} for further details on additional arguments.
}
