% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap_project.R
\name{ody_add_to_datasets}
\alias{ody_add_to_datasets}
\title{Declare an object as belonging to datasets list}
\usage{
ody_add_to_datasets(object, description = NULL)
}
\arguments{
\item{object}{Object to add to datasets. Usually a data frame.}

\item{description}{Optional description of the object.}
}
\description{
Declare an object as belonging to datasets list
}
