% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_utils.R
\name{ody_apply_on_pattern}
\alias{ody_apply_on_pattern}
\title{Apply Function on Pattern}
\usage{
ody_apply_on_pattern(df, fn, pattern, all_any = any, exclude = NULL)
}
\arguments{
\item{df}{A data frame.}

\item{fn}{A function to apply to the columns that match the pattern.}

\item{pattern}{A regular expression pattern to match.}

\item{all_any}{A function (\code{all} or \code{any}) to determine whether all or any of the elements of a column should match the pattern (default is \code{any}).}

\item{exclude}{A character vector of column names to exclude from the pattern matching. That means that the function will not be applied to these columns even if they match the pattern.}
}
\value{
A data frame like the original input with the function applied to the matching columns.
}
\description{
This function applies a specified function to columns of a data frame that match a given pattern.
}
