% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_utils.R
\name{ody_options}
\alias{ody_options}
\title{Extra options for some odytools functions}
\usage{
ody_options(
  label_size = 1,
  minwidth_var = 200,
  n_dec = 1,
  minwidth_level = 100,
  width_density_plot = 650,
  width_bar = 100,
  full_group_label = FALSE
)
}
\arguments{
\item{label_size}{ody_summarise_df: If the variable is labelled, ody_summarise_df shows the label below the var name. This argument controls its relative size.}

\item{minwidth_var}{ody_summarise_df: Minimum width of the Variable column.}

\item{n_dec}{ody_summarise_df: Number of decimals shown in a continuous variable description.}

\item{minwidth_level}{ody_summarise_df: Minimum width of the Level column in the details of a discrete variable.}

\item{width_density_plot}{ody_summarise_df: Total width of the details of continuous variable.}

\item{width_bar}{ody_summarise_df: Width of the percentage bars.}

\item{full_group_label}{ody_summarise_df: Add or not the grouping var name to the group name.}
}
\value{
An list of arguments internaly used in some odytools functions
}
\description{
Extra options for some odytools functions
}
