% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdx.R
\name{ody_pdx_model_sensitivity}
\alias{ody_pdx_model_sensitivity}
\title{(Generalized) Lineal Mixed Model to test PDX sensitivity}
\usage{
ody_pdx_model_sensitivity(data_frame, file_name = NULL, file_dir = getwd())
}
\arguments{
\item{data_frame}{A data frame where:
\itemize{
\item The first column contains the PDX id's.
\item The second column is a two levels factors for sensitivity.
\item The third column is a two levels factor for treatment.
\item The fourth column is the response variable (integer from 0 to 100).
}}

\item{file_name}{Name of the html file. If NULL, the file is named with the date the analysis was performed and the name of the dependent variable.}

\item{file_dir}{Directory where the result is stored. The default is the current working directory.}
}
\value{
A HTML report
}
\description{
Runs a binomial Genralized Mixed Model to test wether the response to a treatment depends on the sensitivity. The response variable is assumed to be a count of "positive" findings over a total count of 100.
}
