% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores_ct.R
\name{ody_qlq_c30_v3}
\alias{ody_qlq_c30_v3}
\title{Calculate QLQ-C30 Version 3 Scores}
\usage{
ody_qlq_c30_v3(
  qlq_c30_df,
  items_index = NULL,
  completeness = c("half", "all", "none")
)
}
\arguments{
\item{qlq_c30_df}{A dataframe containing the QLQ-C30 questionnaire responses.}

\item{items_index}{A numeric vector specifying the column indexes of the items
in the QLQ-C30 questionnaire. If NULL, the last 30 columns are assumed to be the items sorted as they are presented in the questionnaire.}

\item{completeness}{A character vector specifying the completeness criteria:
\itemize{
\item "half" (default): At least half of the items must be available to calculate the score.
\item "all": All items must be available to calculate the score.
\item "none": The score is always calculated as long as at least one item is available.
}}
}
\value{
A dataframe with the calculated scores for each scale (see below) in the
QLQ-C30 questionnaire,binded with any extra column. The item mapping, a
relation between the item number, the scale it belongs to and the column
name in the input data frame, is stored as an attribute called "item_mapping".
}
\description{
This function takes a QLQ-C30 Version 3 questionnaire dataset and calculates
the scores for each scale according to the predefined scoring rules. The dataset
should have at least 30 columns representing the questionnaire items.
}
\section{Global Health Status}{


A high score represent a high quality of life (the HIGHER the BETTER).
\itemize{
\item QL2: Global health status / Quality of life.
}
}

\section{Functional scales}{


High scores represent a high level of functioning (the HIGHER the BETTER)
\itemize{
\item PF2: Physical functioning.
\item RF2: Role functioning.
\item EF: Emotional functioning.
\item CF: Cognitive functioning.
\item SF: Social functioning.
}
}

\section{Symptom scales}{


High scores represent a high level of symptoms/problems (the HIGHER the WORSE).
\itemize{
\item FA: Fatigue.
\item NV: Nausea and vomiting.
\item PA: Pain.
\item DY: Dyspnoea.
\item SL: Insomnia.
\item AP: Appetite loss.
\item CO: Constipation.
\item DI: Diarrhoea.
\item FI: Financial difficulties.
}
}

