% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores_ct.R
\name{ody_qlq_c30_v3}
\alias{ody_qlq_c30_v3}
\title{Calculate QLQ-C30 Version 3 Scores}
\usage{
ody_qlq_c30_v3(qlq_c30, id_col = 1, na_rm = FALSE)
}
\arguments{
\item{qlq_c30}{A dataframe containing the QLQ-C30 questionnaire responses.
This dataframe must contain exactly 30 or 31 columns. If 31
columns are present, one column is treated as an identifier column.}

\item{id_col}{An optional parameter specifying the column index of the
identifier column in \code{qlq_c30} dataframe. The default index is 1.
This parameter is used only when \code{qlq_c30} contains 31 columns.}

\item{na_rm}{A logical value indicating whether NA values should be removed before
calculating the scores. The default is FALSE, so all items needed
to calculate a specific score must be non-NA.}
}
\value{
A dataframe with the calculated scores for each scale in the
QLQ-C30 questionnaire, optionally binded with the identifier column.
}
\description{
This function takes a QLQ-C30 Version 3 questionnaire dataset and calculates
the scores for each scale according to predefined scoring rules. The dataset
should have 30 item columns representing questionnaire items and optionally
an identifier column.
}
