% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap.R
\name{ody_rc_add_label}
\alias{ody_rc_add_label}
\title{Adds labels to a data frame using REDCap metadata.}
\usage{
ody_rc_add_label(df, metadata = NULL, modify_names = FALSE)
}
\arguments{
\item{df}{A data frame to which the labels will be added.}

\item{metadata}{Either the metadata of a REDCap project or a \code{redcap_data} object imported by \code{ody_rc_import}. In the latter case, the metadata will be extracted from the \code{metadata} attribute. If NULL, the function will look for a \code{redcap_data} object in the global environment.}

\item{modify_names}{Logical. If TRUE, the function will modify the names of
the data frame instead of adding labels.}
}
\value{
A data frame with variable labels applied from the REDCap metadata.
}
\description{
This function takes a data frame and applies variable labels derived from
the REDCap metadata column \code{field_label}.
}
