% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap.R
\name{ody_rc_add_site}
\alias{ody_rc_add_site}
\title{Add the sites to a RedCap table}
\usage{
ody_rc_add_site(tbl, rc_data = redcap_data, position = 1)
}
\arguments{
\item{tbl}{The table to add the site to.}

\item{rc_data}{The redcap_data object with the sites information as attribute
(the dag attribute of an ody_rc_import output). Default is redcap_data.}

\item{position}{The position of the site column. Default is 1.}
}
\value{
The same tbl with the site column added.
}
\description{
Add the sites to a RedCap table
}
