% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap_project.R
\name{ody_rc_current}
\alias{ody_rc_current}
\title{Get the name and current import of the project}
\usage{
ody_rc_current(redcap_data)
}
\arguments{
\item{redcap_data}{ody_rc_import import}
}
\description{
Get the name and current import of the project
}
