% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap.R
\name{ody_rc_import}
\alias{ody_rc_import}
\title{Import a RedCap project}
\usage{
ody_rc_import(token = NULL, url = "https://redcap.vhio.net/redcap/api/")
}
\arguments{
\item{token}{API token provided by RedCap.}

\item{url}{RedCap server. The default is the server of VHIO.}
}
\value{
A nested tibble
}
\description{
Import a RedCap project
}
