% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap.R
\name{ody_rc_search_ttm}
\alias{ody_rc_search_ttm}
\title{Search Patients with Specific Treatment}
\usage{
ody_rc_search_ttm(
  rc_data,
  ttm_line,
  ttm_pattern,
  variables_of_interest = c("ttm_startdate", "ttm_enddate", "ttm_pddate",
    "ttm_bestresponse"),
  join_rc_spread = TRUE
)
}
\arguments{
\item{rc_data}{A RedCap data object imported via \code{ody_rc_import}.}

\item{ttm_line}{A vector specifying the treatment line numbers to filter.}

\item{ttm_pattern}{A pattern used to search treatment names.}

\item{variables_of_interest}{A character vector of additional variables to include in the output. Default variables are  "ttm_startdate", "ttm_enddate", "ttm_pddate" and "ttm_bestresponse".}

\item{join_rc_spread}{Logical. If TRUE, joins the filtered cases with the wider RedCap data spread. Defaults to TRUE.}
}
\value{
A tibble with the filtered TTM cases, optionally combined with the wider RedCap data.
}
\description{
Filters RedCap data to identify patients who have received a specific treatment in designated lines, suited for Master-like RedCap projects.
}
