% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap.R
\name{ody_rc_select}
\alias{ody_rc_select}
\title{Title}
\usage{
ody_rc_select(rc_data, ...)
}
\arguments{
\item{rc_data}{RedCap data imported with ody_rc_import.}

\item{...}{Variable names to select.}
}
\value{
A tibble with the selected variables.
}
\description{
Title
}
