% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap.R
\name{ody_rc_select}
\alias{ody_rc_select}
\title{Select variables from a RedCap import}
\usage{
ody_rc_select(
  rc_data,
  ...,
  .is_vector = FALSE,
  .if_different_forms = c("list", "join"),
  .include_aux = FALSE,
  .accept_form_name = TRUE
)
}
\arguments{
\item{rc_data}{RedCap data imported with ody_rc_import.}

\item{...}{Variable names to select. If the name of a form is provided, all the variables belonging to that form will be selected.}

\item{.is_vector}{Logical. If TRUE, the first element of ... is considered a character vector with the names of the variables to be selected.}

\item{.if_different_forms}{What action take if the selected variables belong to different forms.
\itemize{
\item list: It returns a list with an element for each form so only variables belonging to the same form are joined in the same data frame.
\item join: Join all variables creating artifact NAs.
}}

\item{.include_aux}{When a form name is provided, all auxiliary checkbox variables will be added if .include_aux = TRUE}

\item{.accept_form_name}{Logical. If TRUE (default), a form name can be provided in ... to select all variables of that form. Set to FALSE if a variable is named as a form and you need to select the variable instead of the form. For selecting complete forms, it is now recommended to use \code{ody_rc_select_form}.}
}
\value{
A tibble with the selected variables.
}
\description{
Select variables from a RedCap import
}
