% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap.R
\name{ody_rc_select}
\alias{ody_rc_select}
\title{Select variables from a RedCap import}
\usage{
ody_rc_select(rc_data, ..., .include_aux = FALSE)
}
\arguments{
\item{rc_data}{RedCap data imported with ody_rc_import.}

\item{...}{Variable names to select. If the name of a form is provided, all the variables belonguing to that form will be selected.}

\item{.include_aux}{When a form name is provided, all auxiliar checkbox variables will be added if .include_aux = TRUE}
}
\value{
A tibble with the selected variables.
}
\description{
Select variables from a RedCap import
}
