% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap.R
\name{ody_rc_spread}
\alias{ody_rc_spread}
\title{Spread a Classic REDCap project into a 2D table}
\usage{
ody_rc_spread(rc_data = NULL)
}
\arguments{
\item{rc_data}{The object to spread.}
}
\value{
A tibble
}
\description{
The function ody_rc_spread takes a classic project (with no events) and spreads it into a tibble with one row per subject. This is useful for creating Excel exports.
}
\details{
If no data provided, the function checks whether there is a redcap_data object in the environment.
}
