% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap.R
\name{ody_rc_view}
\alias{ody_rc_view}
\title{View a RedCap project}
\usage{
ody_rc_view(data_app = NULL)
}
\arguments{
\item{data_app}{Imported data by ody_rc_import (must be labelled and nested). If no data provided, the function calls ody_rc_import to download it from RedCap.}
}
\value{
An html viewer
}
\description{
View a RedCap project
}
