% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap.R
\name{ody_rc_view}
\alias{ody_rc_view}
\title{View a RedCap project}
\usage{
ody_rc_view(data_app = get("redcap_data"), raw = FALSE)
}
\arguments{
\item{data_app}{Imported data by ody_rc_import (must be labelled and nested).}

\item{raw}{Logical. Show raw data with no labels?}
}
\value{
An html viewer
}
\description{
View a RedCap project
}
