% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap_project.R
\name{ody_read_data}
\alias{ody_read_data}
\title{Read a data file from the project's data folder}
\usage{
ody_read_data(data_file, guess_cols = FALSE)
}
\arguments{
\item{data_file}{Name of the data file to read.}

\item{guess_cols}{Logical; if TRUE, column types are guessed, otherwise
all columns are read as text.}
}
\value{
A data frame containing the contents of the specified file.
}
\description{
Read a data file from the project's data folder
}
\examples{
\dontrun{
  df <- ody_read_data("mydata.csv")
  df <- ody_read_data("mydata.xlsx", guess_cols = TRUE)
}
}
