% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style_methods.R
\name{ody_style}
\alias{ody_style}
\title{Apply "Odytools" styling to an object}
\usage{
ody_style(x)
}
\arguments{
\item{x}{An object usually a graphic or table.}
}
\value{
The same object with the "Odytools" style applied. If the object is not supported, the object is returned as is.
}
\description{
Adds the default style of "Odytools" to some objects.
}
