% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_tables.R
\name{ody_summarise_df}
\alias{ody_summarise_df}
\title{ody_summarise_df}
\usage{
ody_summarise_df(
  data_frame,
  grouping_var = NULL,
  conditions_list = NULL,
  exclude = NULL,
  compare_groups = FALSE,
  use_NA = c("no", "ifany", "always")
)
}
\arguments{
\item{data_frame}{data frame}

\item{grouping_var}{A grouping variable}

\item{conditions_list}{Conditions list to filter variables}

\item{exclude}{Variables that should be excluded}

\item{compare_groups}{Compare groups?}

\item{use_NA}{Add missing values in the report?}
}
\value{
A list
}
\description{
ody_summarise_df
}
\details{
If grouping_var exists and compare_groups = TRUE, the code runs compareGroups::compareGroups(grouping_var ~ variable).
}
