% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_summary.R
\name{ody_summarise_df}
\alias{ody_summarise_df}
\title{ody_summarise_df}
\usage{
ody_summarise_df(
  data_frame,
  grouping_var = NULL,
  conditions_list = NULL,
  exclude = NULL,
  compare_groups = c("no", "p_value", "q_value", "both"),
  use_NA = c("no", "ifany", "always"),
  min_distinct = 5,
  add_overall = TRUE,
  raw_summary = FALSE,
  show_completeness = TRUE,
  show_conditions = FALSE,
  opt_reactable = ody_options(),
  ...
)
}
\arguments{
\item{data_frame}{data frame}

\item{grouping_var}{A grouping variable}

\item{conditions_list}{Conditions list to filter variables}

\item{exclude}{Variables that should be excluded}

\item{compare_groups}{Compare groups?}

\item{use_NA}{Add missing values in the report?}

\item{min_distinct}{Minimal number of distinct cases in a numeric variable to be described as numeric. If the number of distinct cases is lower, the variable is described as it was a factor.}

\item{add_overall}{If TRUE and grouping_var is not NULL, an overall description is added to the detail table of each variable.}

\item{raw_summary}{If TRUE, the function returns a raw summary instead of the defaiult reactable report.}

\item{show_completeness}{Logical, it TRUE it adds completeness information to the main table.}

\item{show_conditions}{If TRUE, the filtering conditions defined by conditions_list ( if any) appear at each filtered variable.}

\item{opt_reactable}{Reactable options. A call to ody_options}

\item{...}{Further aditional options to be directly passed to reactable}
}
\value{
A list
}
\description{
ody_summarise_df
}
\details{
If grouping_var exists and compare_groups = TRUE, the code runs gtsummary::tbl_summary and extracts the p-values.
}
